+++ src/copy.rs
@@ -34,7 +34,7 @@ pub fn copy(input: TokenStream) -> TokenStream {
     let input = parse_macro_input!(input as DeriveInput);
 
     // Extract the generics & fmts for the general impl
-    let generics = match extract_generics("copy", &input.attrs, &input, &Path {
+    let (impl_gen, ty_gen, where_clause) = match extract_generics("copy", &input.attrs, &input, &Path {
         leading_colon: Some(Default::default()),
         segments:      {
             let mut segments = Punctuated::new();
@@ -50,7 +50,6 @@ pub fn copy(input: TokenStream) -> TokenStream {
 
     // Done, build the impl
     let name = &input.ident;
-    let (impl_gen, ty_gen, where_clause) = generics.split_for_impl();
     quote! {
         impl #impl_gen ::std::marker::Copy for #name #ty_gen #where_clause {}
     }