+++ src/debug.rs
@@ -156,7 +156,7 @@ pub fn debug(input: TokenStream) -> TokenStream {
     }
 
     // Extract the generics & fmts for the general impl
-    let generics = match extract_generics("debug", &input.attrs, &input, &Path {
+    let (impl_gen, ty_gen, where_clause) = match extract_generics("debug", &input.attrs, &input, &Path {
         leading_colon: Some(Default::default()),
         segments:      {
             let mut segments = Punctuated::new();
@@ -173,7 +173,6 @@ pub fn debug(input: TokenStream) -> TokenStream {
 
     // Done, build the impl
     let name = &input.ident;
-    let (impl_gen, ty_gen, where_clause) = generics.split_for_impl();
     quote! {
         impl #impl_gen ::std::fmt::Debug for #name #ty_gen #where_clause {
             #[inline]