+++ src/hash.rs
@@ -150,7 +150,7 @@ pub fn hash(input: TokenStream) -> TokenStream {
     }
 
     // Extract the generics & fmts for the general impl
-    let generics = match extract_generics("hash", &input.attrs, &input, &Path {
+    let (impl_gen, ty_gen, where_clause) = match extract_generics("hash", &input.attrs, &input, &Path {
         leading_colon: Some(Default::default()),
         segments:      {
             let mut segments = Punctuated::new();
@@ -167,7 +167,6 @@ pub fn hash(input: TokenStream) -> TokenStream {
 
     // Done, build the impl
     let name = &input.ident;
-    let (impl_gen, ty_gen, where_clause) = generics.split_for_impl();
     quote! {
         impl #impl_gen ::std::hash::Hash for #name #ty_gen #where_clause {
             #[inline]