+++ src/ord.rs
@@ -51,7 +51,7 @@ pub fn ord(input: TokenStream) -> TokenStream {
     let input = parse_macro_input!(input as DeriveInput);
 
     // Extract the generics & fmts for the general impl
-    let generics = match extract_generics("ord", &input.attrs, &input, &Path {
+    let (impl_gen, ty_gen, where_clause) = match extract_generics("ord", &input.attrs, &input, &Path {
         leading_colon: Some(Default::default()),
         segments:      {
             let mut segments = Punctuated::new();
@@ -68,7 +68,6 @@ pub fn ord(input: TokenStream) -> TokenStream {
 
     // Done, build the impl
     let name = &input.ident;
-    let (impl_gen, ty_gen, where_clause) = generics.split_for_impl();
     quote! {
         impl #impl_gen ::std::cmp::Ord for #name #ty_gen #where_clause {
             #[inline]