+++ src/serialize.rs
@@ -174,7 +174,7 @@ pub fn serialize(input: TokenStream) -> TokenStream {
     }
 
     // Extract the generics & fmts for the general impl
-    let generics = match extract_generics("serialize", &input.attrs, &input, &Path {
+    let (impl_gen, ty_gen, where_clause) = match extract_generics("serialize", &input.attrs, &input, &Path {
         leading_colon: Some(Default::default()),
         segments:      {
             let mut segments = Punctuated::new();
@@ -190,7 +190,6 @@ pub fn serialize(input: TokenStream) -> TokenStream {
 
     // Done, build the impl
     let name = &input.ident;
-    let (impl_gen, ty_gen, where_clause) = generics.split_for_impl();
     quote! {
         impl #impl_gen ::serde::Serialize for #name #ty_gen #where_clause {
             fn serialize<SE>(&self, __serializer: SE) -> ::std::result::Result<SE::Ok, SE::Error>